# API Documentation/Core.swagger_collection.yaml
openapi: 3.0.1
info:
  title: Deep Lynx
  description: |-
    The construction of megaprojects has consistently demonstrated challenges for project managers in regard to meeting cost, schedule, and performance requirements. Megaproject construction challenges are common place within megaprojects with many active projects in the United States failing to meet cost and schedule efforts by significant margins. Currently, engineering teams operate in siloed tools and disparate teams where connections across design, procurement, and construction systems are translated manually or over brittle point-to-point integrations. The manual nature of data exchange increases the risk of silent errors in the reactor design, with each silent error cascading across the design. These cascading errors lead to uncontrollable risk during construction, resulting in significant delays and cost overruns. Deep Lynx allows for an integrated platform during design and operations of mega projects.

    The Deep Lynx Core API delivers a few main features.

    1. Provides a set of methods and endpoints for manipulating data in an object oriented database. This allows us to store complex datatypes as records and then to compile them into actual, modifiable objects at run-time. Users can store taxonomies or ontologies in a readable format.

...
  contact: {}
  version: '1.0'
  license:
    name: MIT
    url: 'https://github.com/idaholab/Deep-Lynx/blob/master/LICENSE'
servers:
  - url: '{url}:{port}'
...
      url:
...
security:
  - BearerAuth: []
paths:
...
  '/containers/{container_id}/import/datasources/{data_source_id}/files':
    post:
      tags:
        - Data Sources
      summary: Upload File
      description: 'Uploads a file and metadata to Deep Lynx. This endpoint will accept multiple files and multiple metadata properties as form values. If metadata (additional key value pairs) are provided, an import is created for the data source in addition to the file upload. Transformations can be applied to this import to create nodes and edges in the graph with metadata for the uploaded file(s).'
      operationId: UploadFile
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: data_source_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                oneOf:
                  - properties:
                      value:
                        $ref: '#/components/schemas/FileResponse'
                      isError:
                        type: boolean
                  - properties:
                      value:
                        $ref: '#/components/schemas/FileInfoResponse'
                      isError:
                        type: boolean
                type: object
      deprecated: false
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
  '/containers/{container_id}/files/{file_id}/download':
    get:
      tags:
        - Data Sources
      summary: Download File
      description: Downloads a previously uploaded file.
      operationId: DownloadFile
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content: {}
      deprecated: false
  '/containers/{container_id}/files/{file_id}':
    get:
      tags:
        - Data Sources
      summary: Retrieve File
      description: Get information about a file by ID.
      operationId: RetrieveFile
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    $ref: '#/components/schemas/FileInfoResponse'
                  isError:
                    type: boolean
      deprecated: false
  '/containers/{container_id}/import/datasources/{data_source_id}/imports':
    get:
      tags:
        - Data Sources
      summary: List Imports for Data Source
      operationId: ListImportsForDataSource
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: data_source_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    type: array
                    items:
                      $ref: '#/components/schemas/DataSourceImportResponse'
                  isError:
                    type: boolean
      deprecated: false
      description: List the imports for the datasource.
    post:
      tags:
        - Data Sources
      summary: Create Manual Import
      description: Create a manual import.
      operationId: CreateManualImport
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: data_source_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                test: test
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    $ref: '#/components/schemas/CreateManualImportResponse'
                  isError:
                    type: boolean
      deprecated: false
      x-codegen-request-body-name: Body
...
  '/containers/{container_id}/import/imports/{import_id}/data':
    get:
      tags:
        - Imports
      summary: List Import's Data
      operationId: ListImportsData
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: import_id
          in: path
          required: true
          schema:
            type: string
        - schema:
            type: string
          in: query
          name: count
          description: boolean indicating if the return value should be a count only
        - schema:
            type: integer
            minimum: 0
          in: query
          name: limit
        - schema:
            type: integer
            minimum: 0
          in: query
          name: offset
        - schema:
            type: string
          in: query
          name: sortBy
          description: column to sort results by
        - schema:
            type: boolean
          in: query
          name: sortDesc
          description: boolean indicating if results should be in descending order
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    $ref: '#/components/schemas/ImportResponse'
                  isError:
                    type: boolean
      deprecated: false
      description: List the data for an import.
  '/containers/{container_id}/import/imports/{import_id}':
    delete:
      tags:
        - Imports
      summary: Delete Import
      description: Delete import will delete an import ONLY IF the import has not been processed.
      operationId: DeleteImport
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: import_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content: {}
      deprecated: false
  '/containers/{container_id}/import/imports/{import_id}/data/{data_id}':
    get:
      tags:
        - Imports
      summary: Retrieve Import Data
      description: Retrieve a single piece of data from an import.
      operationId: RetrieveImportData
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: import_id
          in: path
          required: true
          schema:
            type: string
        - name: data_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    $ref: '#/components/schemas/ImportResponse'
                  isError:
                    type: boolean
      deprecated: false
    delete:
      tags:
        - Imports
      summary: Delete Import Data
      description: Delete a single piece of data from an import.
      operationId: DeleteImportData
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: import_id
          in: path
          required: true
          schema:
            type: string
        - name: data_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generic200'
      deprecated: false
    put:
      tags:
        - Imports
      summary: Update Import Data
      operationId: UpdateImportData
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: string
        - name: import_id
          in: path
          required: true
          schema:
            type: string
        - name: data_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    $ref: '#/components/schemas/ImportResponse'
                  isError:
                    type: boolean
      description: Update the data of an existing import.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataStaging'
...
  '/containers/{container_id}/datasources/{source_id}/imports':
    parameters:
      - schema:
          type: string
        name: container_id
        in: path
        required: true
      - schema:
          type: string
        name: source_id
        in: path
        required: true
    post:
      tags:
        - Imports
      summary: Create Import
      operationId: CreateImport
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    $ref: '#/components/schemas/CreateManualImportResponse'
                  isError:
                    type: boolean
      description: Creates a new import.
      requestBody:
        content:
          application/json:
            schema:
              description: ''
              type: object
              x-examples:
                example-1:
                  data_source_id: 1294f879-7f15-40e7-86f4-56177108ba23
                  reference: testing upload
              properties:
                data_source_id:
                  type: string
                  minLength: 1
                reference:
                  type: string
                  minLength: 1
                status_message:
                  type: string
              required:
                - data_source_id
        description: ''
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Generic200:
      description: ''
      type: object
      properties:
        value:
          type: boolean
        isError:
          type: boolean
      required:
        - value
        - isError
      x-examples:
        Generic200Response:
          value:
            value: true
            isError: false
          isError: false
...
    DataStaging:
      description: ''
      type: object
      x-examples:
        example-1:
          data_source_id: dataSourceID
          import_id: importRecordID
          data: {}
          shape_hash: test_hash
      properties:
        data_source_id:
          type: string
          minLength: 1
        import_id:
          type: string
          minLength: 1
        data:
          type: object
        shape_hash:
          type: string
          minLength: 1
      required:
        - data_source_id
        - import_id
        - data
...